+++ pkg/alertmanager/validation/v1alpha1/validation.go
@@ -42,7 +42,7 @@ func ValidateAlertmanagerConfig(amc *monitoringv1alpha1.AlertmanagerConfig) erro
 		return err
 	}
 
-	return validateAlertManagerRoutes(amc.Spec.Route, receivers, muteTimeIntervals, true)
+	return validateRoute(amc.Spec.Route, receivers, muteTimeIntervals, true)
 }
 
 func validateReceivers(receivers []monitoringv1alpha1.Receiver) (map[string]struct{}, error) {
@@ -345,10 +345,11 @@ func validateMSTeamsConfigs(configs []monitoringv1alpha1.MSTeamsConfig) error {
 	return nil
 }
 
-// validateAlertManagerRoutes verifies that the given route and all its children are semantically valid.
-// because of the self-referential issues mentioned in https://github.com/kubernetes/kubernetes/issues/62872
-// it is not currently possible to apply OpenAPI validation to a v1alpha1.Route.
-func validateAlertManagerRoutes(r *monitoringv1alpha1.Route, receivers, muteTimeIntervals map[string]struct{}, topLevelRoute bool) error {
+// validateRoute verifies that the given route and all its children are
+// semantically valid.  because of the self-referential issues mentioned in
+// https://github.com/kubernetes/kubernetes/issues/62872 it is not currently
+// possible to apply OpenAPI validation to a v1alpha1.Route.
+func validateRoute(r *monitoringv1alpha1.Route, receivers, muteTimeIntervals map[string]struct{}, topLevelRoute bool) error {
 	if r == nil {
 		return nil
 	}
@@ -388,7 +389,6 @@ func validateAlertManagerRoutes(r *monitoringv1alpha1.Route, receivers, muteTime
 		}
 	}
 
-	// validate that if defaults are set, they match regex
 	if r.GroupInterval != "" && !durationRe.MatchString(r.GroupInterval) {
 		return fmt.Errorf("groupInterval %s does not match required regex: %s", r.GroupInterval, durationRe.String())
 
@@ -401,13 +401,20 @@ func validateAlertManagerRoutes(r *monitoringv1alpha1.Route, receivers, muteTime
 		return fmt.Errorf("repeatInterval %s does not match required regex: %s", r.RepeatInterval, durationRe.String())
 	}
 
+	for i, m := range r.Matchers {
+		if err := m.Validate(); err != nil {
+			return fmt.Errorf("matcher[%d]: %w", i, err)
+		}
+	}
+
+	// Unmarshal the child routes and validate them recursively.
 	children, err := r.ChildRoutes()
 	if err != nil {
 		return err
 	}
 
 	for i := range children {
-		if err := validateAlertManagerRoutes(&children[i], receivers, muteTimeIntervals, false); err != nil {
+		if err := validateRoute(&children[i], receivers, muteTimeIntervals, false); err != nil {
 			return fmt.Errorf("route[%d]: %w", i, err)
 		}
 	}